Introduce wasm_utils::api_serialization add validation.rs to it and use core_types in it
