+++ core/src/nucleus/actions/validate.rs
@@ -6,7 +6,7 @@ use futures::{future, Async, Future};
 use holochain_core_types::{
     cas::content::AddressableContent, entry::Entry, entry_type::EntryType, hash::HashString,
 };
-use holochain_wasm_utils::validation::ValidationData;
+use holochain_wasm_utils::api_serialization::validation::ValidationData;
 use nucleus::ribosome::callback::{self, CallbackResult};
 use snowflake;
 use std::{sync::Arc, thread};
