+++ core/src/nucleus/ribosome/callback/validate_entry.rs
@@ -2,7 +2,7 @@ extern crate serde_json;
 use context::Context;
 use holochain_core_types::{entry::Entry, entry_type::EntryType, error::HolochainError};
 use holochain_dna::wasm::DnaWasm;
-use holochain_wasm_utils::validation::ValidationData;
+use holochain_wasm_utils::api_serialization::validation::ValidationData;
 use nucleus::{
     ribosome::{
         self,
