+++ wasm_utils/Cargo.toml
@@ -7,4 +7,4 @@ authors = ["Holochain Core Dev Team <dev.support@holo.host>"]
 serde="1"
 serde_derive="1"
 serde_json="1"
-
+holochain_core_types={ path = "../core_types" }
