+++ wasm_utils/src/lib.rs
@@ -5,8 +5,9 @@ extern crate serde;
 extern crate serde_json;
 #[macro_use]
 extern crate serde_derive;
+extern crate holochain_core_types;
 
+pub mod api_serialization;
 pub mod error;
 pub mod memory_allocation;
 pub mod memory_serialization;
-pub mod validation;
